+++ OpenAI-DotNet/Audio/ChunkingStrategy.cs
@@ -7,7 +7,10 @@ namespace OpenAI.Audio
     public sealed class ChunkingStrategy
     {
         /// <summary>
-        /// 
+        /// Controls how the audio is cut into chunks.
+        /// When set to "auto", the server first normalizes loudness and then uses voice activity detection (VAD) to choose boundaries.
+        /// server_vad object can be provided to tweak VAD detection parameters manually.
+        /// If unset, the audio is transcribed as a single block.
         /// </summary>
         /// <param name="prefixPaddingMs">
         /// Amount of audio to include before the VAD detected speech (in milliseconds).
