+++ OpenAI-DotNet/Realtime/DisabledVAD.cs
@@ -1,0 +1,11 @@
+﻿namespace OpenAI.Realtime
+{
+    public sealed class DisabledVAD : IVoiceActivityDetectionSettings
+    {
+        public TurnDetectionType Type => TurnDetectionType.Disabled;
+
+        public bool CreateResponse => false;
+
+        public bool InterruptResponse => false;
+    }
+}
