+++ OpenAI-DotNet/Realtime/IVoiceActivityDetectionSettings.cs
@@ -1,0 +1,9 @@
+﻿namespace OpenAI.Realtime
+{
+    public interface IVoiceActivityDetectionSettings
+    {
+        public TurnDetectionType Type { get; }
+        public bool CreateResponse { get; }
+        public bool InterruptResponse { get; }
+    }
+}
